make the keys build with s3_client compatible with minio on Windows
In ```s3_client.py``` and ```s3_boto_data_store.py```, Python's ```os.path.normpath``` was used to build the keys for unload/download files to/from minio server. However, it will use ```\\``` to replace ```/``` in the keys, which in incompatible with minio's way of path to files. So, I simply skipped ```os.path.normpath``` and build the key strings direstly with ```/```.